クッキー認証を別ファイルにした。
